introduce GenericSystemRunner module with system_with_output method, remove system_with_output from nashorn, nodejs and strictnodejs runners, include GenericSystemRunner in nashorn, nodejs and strictnodejs runners
